Create NormalityToAbnormalReturn
The goal of this project is to calculate the monthly average raw returns and the abnormal return using the Fama-French three-factor model for all 11 skewness-based and kurtosis-based portfolios. The data used are mainboard and GEM board data from January 2015 to May 6th, 2022 taken from the Chinese stock market trading.

In the case of the portfolios that were ranked by skewness, each portfolio return’s plot all clumped up together and does not give any indication of higher return when it is following normal distribution. A possible explanation to this phenomenon is because there are many factors that affect return, such as the average return (mean) and the volatility of the stock (standard deviation).

On the other hand, the portfolios that were ranked by kurtosis we can imply that the lower the kurtosis is the higher the return is. However, we should note that the lowest ranked company has a kurtosis of 1. So an increase of kurtosis will first get closer to normal distribution, nevertheless, too much increases result in a very high peak than its tails for its distributions of data. Hence, the higher the portfolio, the lower the return because it has a very high peak (kurtosis is too high).