Create Standardized Unexpected Earnings vs Abnormal Return.R
In this part, I analyzed Post Earning Announcement Drift (PEAD), which is the tendency for a stock’s cumulative abnormal returns to drift in the same direction of an earnings surprise for several weeks (even several months) following an earnings announcement.

The graph illustrates the cumulative abnormal returns from 120 days before the announcement of earnings and 120 days after the announcement of earnings. From the graph, it is obvious that the upper ranked portfolios (portfolio 1, 2, and 3) generate higher cumulative abnormal returns compared to the lower ranked (portfolio 8, 9, and 10). The cumulative abnormal returns represent how different the realized return is to the expected return. The first few portfolios have higher cumulative abnormal returns because the realized return exceeds the expected return, while the lower cumulative abnormal returns indicate the realized return falls behind the expected return. Note that the portfolios are ranked and ordered by their Standardized Earnings. Portfolio 1 has the highest standardized earnings, while portfolio 10 has the lowest standardized earnings, on average. This conclusion makes sense because the higher the standardized earnings are for a company in that period, it is highly likely that it will generate a good return in the next period as well. This is because a company with higher standardized earnings is a more desirable investment for investors.